+++ CHANGES.rst
@@ -45,6 +45,9 @@ Bugs fixed
   term indices before accessing them.
 * #11233: linkcheck: match redirect URIs against :confval:`linkcheck_ignore` by
   overriding session-level ``requests.get_redirect_target``.
+* #13195: viewcode: Fix issue where import paths differ from the directory
+  structure.
+  Patch by Ben Egan and Adam Turner.
 
 Testing
 -------